+++ ASSETS/ExtraFunctions/demoFunctions.js
@@ -150,7 +150,7 @@ function playSelectedBumber(selectList){
     data.webplayout    = '-';                                  // ** NO VIDEO IN WEB PLAYER!!!!! For now. **
     data.playoptions   =  'CUT 0 Linear RIGHT'                 // Additional play options
     data.command       = 'play';
-    x = ajaxpost('/gc/controlvideo',data,'true'); // true is prepopulation
+    x = ajaxpost('/gc/controlvideo',data,'true');               // true is prepopulation
     console.log('Response:',x);
 }
 
@@ -159,14 +159,14 @@ function lottieBumber(){
     // request ..... id of select list
     // returns ..... post a server command
     let data = {};
-    data.relpath       = 'smartpx_internal/GC_PACK_2/BUMPER.html';    // template file
-    data.relpathCCG    = data.relpath.split('.htm')[0];		 // videofile (without file extension!)
-    data.playserver    = 'OVERLAY';                          // CCG server or '-'
-    data.playchannel   = '1';                                // Channel
-    data.playlayer     = '20';                               // Layer
+    data.relpath       = 'smartpx_internal/GC_PACK_2/BUMPER.html';      // template file
+    data.relpathCCG    = data.relpath.split('.htm')[0];		            // videofile (without file extension!)
+    data.playserver    = 'OVERLAY';                                     // CCG server or '-'
+    data.playchannel   = '1';                                           // Channel
+    data.playlayer     = '20';                                          // Layer
     data.webplayout    = '20';                               
     data.command       = 'play';
-    x = ajaxpost('/gc/playout',data,'true');                  // true is prepopulation
+    x = ajaxpost('/gc/playout',data,'true');                            // true is prepopulation
     console.log('Response:',x);
 }